Improve ClassCountProcessor
