+++ detekt-metrics/src/main/kotlin/dev/detekt/metrics/processors/ClassCountProcessor.kt
@@ -1,24 +1,29 @@
 package dev.detekt.metrics.processors
 
-import com.intellij.openapi.util.Key
-import dev.detekt.api.DetektVisitor
+import dev.detekt.api.Detektion
+import dev.detekt.api.FileProcessListener
+import dev.detekt.api.ProjectMetric
 import org.jetbrains.kotlin.psi.KtClass
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.psiUtil.collectDescendantsOfType
+import kotlin.concurrent.atomics.AtomicInt
+import kotlin.concurrent.atomics.ExperimentalAtomicApi
+import kotlin.concurrent.atomics.plusAssign
 
-class ClassCountProcessor : AbstractProjectMetricProcessor() {
-
+@OptIn(ExperimentalAtomicApi::class)
+class ClassCountProcessor : FileProcessListener {
+    private val count: AtomicInt = AtomicInt(0)
     override val id: String = "ClassCountProcessor"
-    override val visitor = ClassCountVisitor()
-    override val key = numberOfClassesKey
-}
 
-val numberOfClassesKey = Key<Int>("number of classes")
+    override fun onStart(files: List<KtFile>) {
+        count.store(0)
+    }
 
-class ClassCountVisitor : DetektVisitor() {
+    override fun onProcess(file: KtFile) {
+        count += file.collectDescendantsOfType<KtClass>().size
+    }
 
-    override fun visitKtFile(file: KtFile) {
-        super.visitKtFile(file)
-        file.putUserData(numberOfClassesKey, file.collectDescendantsOfType<KtClass>().size)
+    override fun onFinish(files: List<KtFile>, result: Detektion) {
+        result.add(ProjectMetric("number of classes", count.load()))
     }
 }
