+++ detekt-metrics/src/test/kotlin/dev/detekt/metrics/processors/ClassCountProcessorSpec.kt
@@ -20,4 +20,36 @@ class ClassCountProcessorSpec {
         assertThat(detektion.metrics).singleElement()
             .isEqualTo(ProjectMetric("number of classes", 7))
     }
+
+    @Test
+    fun twoClassesInSeparateFile() {
+        val detektion = ClassCountProcessor().invoke(
+            compileContentForTest(default),
+            compileContentForTest(classWithFields)
+        )
+
+        assertThat(detektion.metrics).singleElement()
+            .isEqualTo(ProjectMetric("number of classes", 2))
+    }
+
+    @Test
+    fun oneClassWithOneNestedClass() {
+        val detektion = ClassCountProcessor().invoke(
+            compileContentForTest(complexClass)
+        )
+
+        assertThat(detektion.metrics).singleElement()
+            .isEqualTo(ProjectMetric("number of classes", 2))
+    }
+
+    @Test
+    fun testEnumAndInterface() {
+        val detektion = ClassCountProcessor().invoke(
+            compileContentForTest(emptyEnum),
+            compileContentForTest(emptyInterface)
+        )
+
+        assertThat(detektion.metrics).singleElement()
+            .isEqualTo(ProjectMetric("number of classes", 2))
+    }
 }
